refactor(sdk): extract resilience policy into session-supervisor.ts
Move the three transport-resilience mechanisms (ICE-blip grace +
network listeners, automatic session re-dial, DC-silence watchdog) and
their tunables out of the ReachyMini class into a dedicated
SessionSupervisor. The supervisor owns the RECOVERY POLICY only; every
side effect goes through a narrow deps interface implemented by the
class as closures over its private state, so the class keeps owning the
transport (pc/dc/pending promises) and `_teardownForRedial` stays with
it. ~530 lines out of reachy-mini.ts, no behavior change; redial and
watchdog tests now drive the supervisor's public surface directly.
